fix: 修复所有 clippy 警告
修复的问题：
- 使用 strip_prefix 替代手动字符串切片 (manual_strip)
- 使用 next_back 替代 last 以提高性能 (double_ended_iterator_last)
- 移除不必要的类型转换 (unnecessary_cast)
- 移除不必要的引用借用 (needless_borrows_for_generic_args)
- 重构 audit_service.rs 的 log_action 方法，引入 AuditLogParams 结构体解决参数过多问题 (too_many_arguments)
- 使用 is_some_and 替代 map_or (unnecessary_map_or)

所有更改已通过 cargo clippy -- -D warnings 检查

+++ src/handlers/metrics.rs
@@ -32,7 +32,7 @@ pub async fn metrics_export(State(state): State<Arc<AppState>>) -> Json<MetricsR
         http_request_duration_p50_ms: 0.0,
         http_request_duration_p95_ms: 0.0,
         http_request_duration_p99_ms: 0.0,
-        db_pool_size: state.db.size() as u32,
+        db_pool_size: state.db.size(),
         db_pool_idle: state.db.num_idle() as u32,
         process_uptime_secs: crate::handlers::health::get_uptime(),
     })
